Addressed PR comments, made clear() a test function and added a test case
